App/requirements.txt
Flask>=1.0,<=1.1.2
azure-storage-blob
azure-identity
connexion
connexion[swagger-ui]
opencensus
applicationinsights